requirements.txt
Click==7.0
decorator==4.4.0
Flask==1.1.1
gunicorn==19.9.0
imageio==2.4.1
itsdangerous==1.1.0
Jinja2==2.10.1
lmdb==0.97
MarkupSafe==1.1.1
networkx==2.3
numpy==1.17.1
opencv-python
Pillow==6.1.0
PyWavelets==1.0.3
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
torch==1.0.1 
torchvision==0.2.2
Werkzeug==0.15.5
